External Zip support (via 7zip or zip) External 7Zip support External TorrentZip support (still incomplete) Merged modes (unfinished, unvalidated) NonMerged modes (need validation)

+++ JRomManager/src/jrm/compressors/SevenZipArchive.java
@@ -1,10 +1,8 @@
 package jrm.compressors;
 
-import java.io.BufferedReader;
-import java.io.Closeable;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStreamReader;
+import java.io.InputStream;
 import java.nio.file.Files;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -15,35 +13,64 @@ import org.apache.commons.io.FileUtils;
 import jrm.misc.FindCmd;
 import jrm.misc.Settings;
 
-public class SevenZipArchive implements Closeable, AutoCloseable
+public class SevenZipArchive implements Archive
 {
-	File tempDir = null;
-	File archive;
+	private File tempDir = null;
+	private File archive;
+	private String cmd;
+
+	private List<String> cmd_add = new ArrayList<>();
 
 	public SevenZipArchive(File archive) throws IOException
 	{
 		this.archive = archive;
-		this.tempDir = Files.createTempDirectory("JRM").toFile();
+		this.cmd = Settings.getProperty("7z_cmd", FindCmd.find7z());
+		if(!new File(this.cmd).exists() && !new File(this.cmd+".exe").exists())
+			throw new IOException(this.cmd+" does not exists");
 	}
 
 	@Override
 	public void close() throws IOException
 	{
-		FileUtils.deleteDirectory(tempDir);
+		if(tempDir != null)
+		{
+			int err = 0;
+			if(cmd_add.size()>0)
+			{
+				err = -1;
+				Process process = new ProcessBuilder(cmd_add).directory(tempDir).start();
+				try
+				{
+					err = process.waitFor();
+				}
+				catch (InterruptedException e)
+				{
+					e.printStackTrace();
+				}
+			}
+			FileUtils.deleteDirectory(tempDir);
+			if(err!=0)
+				throw new IOException("Process returned "+err);
+		}
 	}
 
+	public File getTempDir() throws IOException
+	{
+		if(tempDir == null)
+			this.tempDir = Files.createTempDirectory("JRM").toFile();
+		return tempDir;
+	}
+	
 	public File extract(String entry) throws IOException
 	{
-		List<String> cmd = new ArrayList<>(Collections.singletonList(Settings.getProperty("7z_cmd", FindCmd.find7z())));
-		String args = Settings.getProperty("7z_extract_args", SevenZipExtractOptions.SEVENZIP.toString());
-		String[] argv = args.replace("%1", archive.getAbsolutePath()).replace("%2", entry).replace("%4", tempDir.getAbsolutePath()).split("\\s");
-		Collections.addAll(cmd, argv);
-		ProcessBuilder pb = new ProcessBuilder(cmd);
+		List<String> cmd = new ArrayList<>();
+		Collections.addAll(cmd, Settings.getProperty("7z_cmd", FindCmd.find7z()), "x", "-y", archive.getAbsolutePath(), entry);
+		ProcessBuilder pb = new ProcessBuilder(cmd).directory(getTempDir());
 		Process process = pb.start();
 		try
 		{
 			process.waitFor();
-			File result = new File(tempDir, entry);
+			File result = new File(getTempDir(), entry);
 			if (result.exists())
 				return result;
 		}
@@ -54,13 +81,29 @@ public class SevenZipArchive implements Closeable, AutoCloseable
 		return null;
 	}
 
+	public InputStream extract_stdout(String entry) throws IOException
+	{
+		List<String> cmd = new ArrayList<>();
+		Collections.addAll(cmd, Settings.getProperty("7z_cmd", FindCmd.find7z()), "x", "-y", "-so", archive.getAbsolutePath(), entry);
+		return new ProcessBuilder(cmd).start().getInputStream();
+	}
+
+	public int add(String entry) throws IOException
+	{
+		return add(getTempDir(), entry);
+	}
+
 	public int add(File baseDir, String entry) throws IOException
 	{
-		List<String> cmd = new ArrayList<>(Collections.singletonList(Settings.getProperty("7z_cmd", FindCmd.find7z())));
-		String args = Settings.getProperty("7z_add_args", SevenZipAddOptions.SEVENZIP_ULTRA.toString());
-		String[] argv = args.replace("%1", archive.getAbsolutePath()).replace("%2", entry).replace("%4", baseDir.getAbsolutePath()).split("\\s");
-		Collections.addAll(cmd, argv);
-		ProcessBuilder pb = new ProcessBuilder(cmd);
+		if(cmd_add.size()==0)
+		{
+			Collections.addAll(cmd_add, Settings.getProperty("7z_cmd", FindCmd.find7z()), "a", "-y", "-r");
+			Collections.addAll(cmd_add, Settings.getProperty("7z_args", SevenZipOptions.SEVENZIP_ULTRA.toString()).split("\\s"));
+			cmd_add.add(archive.getAbsolutePath());
+		}
+		cmd_add.add(entry);
+		return 0;
+/*		ProcessBuilder pb = new ProcessBuilder(cmd).directory(baseDir);
 		Process process = pb.start();
 		try
 		{
@@ -70,15 +113,39 @@ public class SevenZipArchive implements Closeable, AutoCloseable
 		{
 			e.printStackTrace();
 		}
-		return -1;
+		return -1;*/
+	}
+
+	public int add_stdin(InputStream src, String entry) throws IOException
+	{
+		FileUtils.copyInputStreamToFile(src, new File(getTempDir(),entry));
+		if(cmd_add.size()==0)
+		{
+			Collections.addAll(cmd_add, this.cmd, "a", "-y", "-r");
+			Collections.addAll(cmd_add, Settings.getProperty("7z_args", SevenZipOptions.SEVENZIP_ULTRA.toString()).split("\\s"));
+			cmd_add.add(archive.getAbsolutePath());
+		}
+		cmd_add.add(entry);
+		return 0;
+/*		ProcessBuilder pb = new ProcessBuilder(cmd).directory(getTempDir());
+		Process process = pb.start();
+		try
+		{
+			return process.waitFor();
+		}
+		catch (InterruptedException e)
+		{
+			e.printStackTrace();
+		}
+		return -1;*/
 	}
 
 	public int delete(String entry) throws IOException
 	{
-		List<String> cmd = new ArrayList<>(Collections.singletonList(Settings.getProperty("7z_cmd", FindCmd.find7z())));
-		String args = Settings.getProperty("7z_del_args", SevenZipDeleteOptions.SEVENZIP_ULTRA.toString());
-		String[] argv = args.replace("%1", archive.getAbsolutePath()).replace("%2", entry).split("\\s");
-		Collections.addAll(cmd, argv);
+		List<String> cmd = new ArrayList<>();
+		Collections.addAll(cmd, Settings.getProperty("7z_cmd", FindCmd.find7z()), "d");
+		Collections.addAll(cmd, Settings.getProperty("7z_args", SevenZipOptions.SEVENZIP_ULTRA.toString()).split("\\s"));
+		Collections.addAll(cmd, archive.getAbsolutePath(), entry);
 		ProcessBuilder pb = new ProcessBuilder(cmd);
 		Process process = pb.start();
 		try
@@ -94,10 +161,10 @@ public class SevenZipArchive implements Closeable, AutoCloseable
 
 	public int rename(String entry, String newname) throws IOException
 	{
-		List<String> cmd = new ArrayList<>(Collections.singletonList(Settings.getProperty("7z_cmd", FindCmd.find7z())));
-		String args = Settings.getProperty("7z_ren_args", SevenZipRenameOptions.SEVENZIP_ULTRA.toString());
-		String[] argv = args.replace("%1", archive.getAbsolutePath()).replace("%2", entry).replace("%3", newname).split("\\s");
-		Collections.addAll(cmd, argv);
+		List<String> cmd = new ArrayList<>();
+		Collections.addAll(cmd, Settings.getProperty("7z_cmd", FindCmd.find7z()), "rn");
+		Collections.addAll(cmd, Settings.getProperty("7z_args", SevenZipOptions.SEVENZIP_ULTRA.toString()).split("\\s"));
+		Collections.addAll(cmd, archive.getAbsolutePath(), entry, newname);
 		ProcessBuilder pb = new ProcessBuilder(cmd);
 		Process process = pb.start();
 		try
